Test only the rolling releases of distros
This role does not depend on any particular features of distros to have
wide platform coverage - we only use ansible's lineinfile module to set
variables up in /etc/environment. We just need a way for this to be
tested on all major distros and to be alerted if any of them intend to
break us.

Furthermore, we quicken up the Travis CI test times.